#239 Fix next.js submit endpoint.
